+++ domain/src/test/java/ru/fastdelivery/domain/common/weight/WeightTest.java
@@ -16,15 +16,15 @@ class WeightTest {
     @Test
     @DisplayName("Попытка создать отрицательный вес -> исключение")
     void whenGramsBelowZero_thenException() {
-        var weightGrams = new BigInteger("-1");
+        var weightGrams = new BigDecimal("-1");
         assertThatThrownBy(() -> new Weight(weightGrams))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     void equalsTypeWidth_same() {
-        var weight = new Weight(new BigInteger("1000"));
-        var weightSame = new Weight(new BigInteger("1000"));
+        var weight = new Weight(new BigDecimal("1000"));
+        var weightSame = new Weight(new BigDecimal("1000"));
 
         assertThat(weight)
                 .isEqualTo(weightSame)
@@ -33,7 +33,7 @@ class WeightTest {
 
     @Test
     void equalsNull_false() {
-        var weight = new Weight(new BigInteger("4"));
+        var weight = new Weight(new BigDecimal("4"));
 
         assertThat(weight).isNotEqualTo(null);
     }
@@ -42,7 +42,7 @@ class WeightTest {
     @CsvSource({ "1000, 1, -1",
             "199, 199, 0",
             "50, 999, 1" })
-    void compareToTest(BigInteger low, BigInteger high, int expected) {
+    void compareToTest(BigDecimal low, BigDecimal high, int expected) {
         var weightLow = new Weight(low);
         var weightHigh = new Weight(high);
 
@@ -53,18 +53,18 @@ class WeightTest {
     @Test
     @DisplayName("Добавление положительного веса -> вес увеличился")
     void whenAddPositiveWeight_thenWeightIsIncreased() {
-        var weightBase = new Weight(new BigInteger("1000"));
-        var actual = weightBase.add(new Weight(new BigInteger("1000")));
+        var weightBase = new Weight(new BigDecimal("1000"));
+        var actual = weightBase.add(new Weight(new BigDecimal("1000")));
 
         assertThat(actual)
-                .isEqualTo(new Weight(new BigInteger("2000")));
+                .isEqualTo(new Weight(new BigDecimal("2000")));
     }
 
     @Test
     @DisplayName("Первый вес больше второго -> true")
     void whenFirstWeightGreaterThanSecond_thenTrue() {
-        var weightBig = new Weight(new BigInteger("1001"));
-        var weightSmall = new Weight(new BigInteger("1000"));
+        var weightBig = new Weight(new BigDecimal("1001"));
+        var weightSmall = new Weight(new BigDecimal("1000"));
 
         assertThat(weightBig.greaterThan(weightSmall)).isTrue();
     }
@@ -72,7 +72,7 @@ class WeightTest {
     @Test
     @DisplayName("Запрос количество кг -> получено корректное значение")
     void whenGetKilograms_thenReceiveKg() {
-        var weight = new Weight(new BigInteger("1001"));
+        var weight = new Weight(new BigDecimal("1001"));
 
         var actual = weight.kilograms();
