+++ web/src/main/java/ru/fastdelivery/presentation/api/request/CargoPackage.java
@@ -2,11 +2,12 @@ package ru.fastdelivery.presentation.api.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public record CargoPackage(
         @Schema(description = "Вес упаковки, граммы", example = "5667.45")
-        BigInteger weight,
+        BigDecimal weight,
         @Schema(description = "Длина упаковки, мм", example = "345")
         BigInteger length,
         @Schema(description = "Ширина упаковки, мм", example = "589")
